Level Design Completed, Pooling System Upgraded, Star and Diamond Collectables Added
Level Designer Completed
4 Level have maden, Under Level Designer Cs, Load and Create Level Buttons working and can adjust levels. Change Finish line, stack count, add diamond or star.
Star and Diamond Collectables Added,
Two collectables added, have particle, static pool and dynamic pool.
Pooling system changed,
Monobehaviour class deleted, it is now normal class, was using construction not setting parameters via new class and function.
Cinemachine Orbital Rotation added,
Added new camera which is rotation around player when success.

+++ Assets/Scripts/ProjectTwo/Core/Scriptable/LevelDataScriptable.cs
@@ -7,7 +7,7 @@ namespace GameTwo
     public class LevelDataScriptable : ScriptableObject
     {
         public int stackCount;
-        public float platformLength;
+        public int platformLength;
         public List<StarData> starData;
         public List<DiamondData> diamondData;
         public LevelDataScriptable(int stackCount, int platformLength,List<StarData> starData, List<DiamondData> diamondData)
